Time limit for tasks

+++ computation_manager.h
@@ -6,8 +6,8 @@
 void CM_Init();
 void CM_Shutdown();
 
-Group* CM_NewGroup(int x, int limit);
-Task* CM_NewTask(int groupIdx, int limit, char compSymbol);
+Group* CM_NewGroup(int x, double limit);
+Task* CM_NewTask(int groupIdx, double limit, char compSymbol);
 void CM_Run();
 void CM_Cancel(int groupIdx, int componentIdx);
 void CM_SetRunning(bool running);
